+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.1] = 2025-12-17
+## [1.0.2] - 2025-12-18
+- Update readme for actions to reflect the latest release version
+- Update actions to create consitent github issue labels
+- Update MCP documentation to specify missing server url
+- Update K8S remediation agent to remove requiring yq and simplify manifest handling
+
+## [1.0.1] - 2025-12-17
 - Minor changes to readme and example workflows
 
 ## [1.0.0] - 2025-12-16